test (module-order, module-table) : order와 table모듈에 대한 기존 테스트 코드 이관 및 테스트코드 작성
