+++ README.md
@@ -32,7 +32,7 @@
 
 **4단계 멀티모듈 적용**
 - [ ] 컨텍스트 간의 독립된 모듈로 분리한다. 
-  - [product], [menugroup], [menu], [order], [table], [tablegroup] 
+  - [product], [menu & menugroup], [menu], [order], [table & tablegroup] 
 
 ---
 ## 요구 사항
