+++ module-order/src/test/java/kitchenpos/ModuleOrderApplicationTests.java
@@ -1,0 +1,10 @@
+package kitchenpos;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+class ModuleOrderApplicationTests {
+    @Test
+    void contextLoads() {}
+}
