+++ module-order/src/test/java/kitchenpos/order/acceptance/code/AcceptanceTest.java
@@ -1,0 +1,71 @@
+package kitchenpos.order.acceptance.code;
+
+import io.restassured.RestAssured;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.web.server.LocalServerPort;
+import org.springframework.http.MediaType;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+public class AcceptanceTest {
+
+    private static final String LOCATION_HEADER_KEY = "Location";
+    private static final String SLASH = "/";
+
+    @LocalServerPort
+    int port;
+
+    @Autowired
+    private DatabaseCleanup databaseCleanup;
+
+    @BeforeEach
+    public void setUp() {
+        RestAssured.port = port;
+        databaseCleanup.execute();
+    }
+
+    public static ExtractableResponse<Response> get(String path) {
+        return RestAssured.given().log().all()
+                .when().get(path)
+                .then().log().all()
+                .extract();
+    }
+
+    public static <T> ExtractableResponse<Response> post(String path, T requestBody) {
+        return RestAssured.given().log().all()
+                .body(requestBody)
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .when().post(path)
+                .then().log().all()
+                .extract();
+    }
+
+    public static <T> ExtractableResponse<Response> put(String path, T requestBody) {
+        return RestAssured.given().log().all()
+                .body(requestBody)
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .when().put(path)
+                .then().log().all()
+                .extract();
+    }
+
+    public static ExtractableResponse<Response> delete(String path) {
+        return RestAssured.given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .when().delete(path)
+                .then().log().all()
+                .extract();
+    }
+
+    public static Long parseIdFromLocationHeader(ExtractableResponse<Response> response) {
+        String id = response.header(LOCATION_HEADER_KEY).split(SLASH)[3];
+        return Long.parseLong(id);
+    }
+
+    public static String parseURIFromLocationHeader(ExtractableResponse<Response> response) {
+        return response.header(LOCATION_HEADER_KEY);
+    }
+}
