+++ module-order/src/test/java/kitchenpos/order/fixture/OrderLineItemFixtureFactory.java
@@ -1,0 +1,13 @@
+package kitchenpos.order.fixture;
+
+
+import kitchenpos.order.domain.OrderLineItem;
+
+public class OrderLineItemFixtureFactory {
+
+    private OrderLineItemFixtureFactory() {}
+
+    public static OrderLineItem create(Long menuId, long quantity) {
+        return OrderLineItem.of(menuId, quantity);
+    }
+}
