+++ module-order/src/test/resources/application.properties
@@ -1,0 +1,8 @@
+logging.level.org.springframework.jdbc.core=TRACE
+logging.level.org.hibernate.SQL = debug
+logging.level.org.hibernate.type.descriptor.sql = trace
+spring.h2.console.enabled=true
+
+spring.jpa.properties.hibernate.show_sql=true
+spring.jpa.properties.hibernate.format_sql=true
+spring.jpa.hibernate.ddl-auto=create
