+++ module-table/src/test/java/kitchenpos/ModuleTableApplicationTests.java
@@ -1,0 +1,10 @@
+package kitchenpos;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+class ModuleTableApplicationTests {
+    @Test
+    void contextLoads() {}
+}
