+++ module-table/src/test/java/kitchenpos/table/fixture/OrderTableFixtureFactory.java
@@ -1,0 +1,13 @@
+package kitchenpos.table.fixture;
+
+
+import kitchenpos.table.domain.OrderTable;
+
+public class OrderTableFixtureFactory {
+
+    private OrderTableFixtureFactory() {}
+
+    public static OrderTable createWithGuest(boolean empty, int numberOfGuest) {
+        return OrderTable.of(empty, numberOfGuest);
+    }
+}
