--- src/test/java/kitchenpos/domain/order/OrderLineItemTest.java
@@ -1,136 +1,0 @@
-package kitchenpos.domain.order;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
-import static org.junit.jupiter.api.Assertions.assertAll;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-
-import com.google.common.collect.Lists;
-import java.math.BigDecimal;
-import kitchenpos.fixture.MenuFixtureFactory;
-import kitchenpos.fixture.MenuGroupFixtureFactory;
-import kitchenpos.fixture.MenuProductFixtureFactory;
-import kitchenpos.fixture.OrderFixtureFactory;
-import kitchenpos.fixture.OrderTableFixtureFactory;
-import kitchenpos.fixture.ProductFixtureFactory;
-import kitchenpos.order.application.OrderValidator;
-import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuProduct;
-import kitchenpos.menu.domain.MenuRepository;
-import kitchenpos.menu.domain.MenuGroup;
-import kitchenpos.menu.domain.MenuGroupRepository;
-import kitchenpos.order.domain.Order;
-import kitchenpos.order.domain.OrderLineItem;
-import kitchenpos.order.domain.OrderRepository;
-import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order.exception.NotEqualsMenuAndOrderLineItemMenuException;
-import kitchenpos.product.domain.Product;
-import kitchenpos.product.domain.ProductRepository;
-import kitchenpos.table.domain.OrderTable;
-import kitchenpos.table.domain.OrderTableRepository;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.FilterType;
-import org.springframework.stereotype.Component;
-
-@DataJpaTest(includeFilters = @ComponentScan.Filter(type = FilterType.ANNOTATION, classes = Component.class))
-class OrderLineItemTest {
-
-    @Autowired
-    private MenuGroupRepository menuGroupRepository;
-
-    @Autowired
-    private ProductRepository productRepository;
-
-    @Autowired
-    private MenuRepository menuRepository;
-
-    @Autowired
-    private OrderRepository orderRepository;
-
-    @Autowired
-    private OrderTableRepository orderTableRepository;
-
-    @Autowired
-    private OrderValidator orderValidator;
-
-    private MenuGroup 메뉴그룹;
-    private Menu 메뉴;
-    private Product 상품_1;
-    private Product 상품_2;
-    private MenuProduct 메뉴_상품_1;
-    private MenuProduct 메뉴_상품_2;
-    private OrderTable 주문테이블;
-    private Order 주문;
-
-    @BeforeEach
-    void setUp() {
-        메뉴그룹 = MenuGroupFixtureFactory.create("메뉴그룹");
-        메뉴그룹 = menuGroupRepository.save(메뉴그룹);
-
-        상품_1 = ProductFixtureFactory.create("상품_1", BigDecimal.valueOf(10_000));
-        상품_2 = ProductFixtureFactory.create("상품_2", BigDecimal.valueOf(20_000));
-        상품_1 = productRepository.save(상품_1);
-        상품_2 = productRepository.save(상품_2);
-
-        메뉴 = MenuFixtureFactory.create("메뉴", BigDecimal.valueOf(30_000), 메뉴그룹.getId());
-        메뉴_상품_1 = MenuProductFixtureFactory.create(1L, 메뉴, 상품_1.getId(), 1L);
-        메뉴_상품_2 = MenuProductFixtureFactory.create(2L, 메뉴, 상품_2.getId(), 2L);
-
-        메뉴_상품_1.mappedByMenu(메뉴);
-        메뉴_상품_2.mappedByMenu(메뉴);
-        메뉴 = menuRepository.save(메뉴);
-
-        주문테이블 = OrderTableFixtureFactory.createWithGuest(false, 2);
-        주문테이블 = orderTableRepository.save(주문테이블);
-
-        주문 = OrderFixtureFactory.create(주문테이블.getId(), OrderStatus.COOKING, Lists.newArrayList());
-        주문 = orderRepository.save(주문);
-    }
-
-    @DisplayName("OrderLineItem 을 생성할 수 있다. (Menu, Quantity, Order)")
-    @Test
-    void create01() {
-        // given & when
-        OrderLineItem 주문항목 = OrderLineItem.of(메뉴.getId(), 1L);
-
-        // then
-        assertAll(
-                () -> assertNotNull(주문항목),
-                () -> assertEquals(1L, 주문항목.findQuantity())
-        );
-    }
-
-    @DisplayName("OrderLineItem 생성 시 Menu가 null이면 예외가 발생한다.")
-    @Test
-    void create02() {
-        // given
-        OrderLineItem orderLineItem = OrderLineItem.of(0L, 1L);
-        Order order = OrderFixtureFactory.create(주문테이블.getId(), OrderStatus.COOKING, Lists.newArrayList(orderLineItem));
-
-        // when & then
-        assertThatExceptionOfType(NotEqualsMenuAndOrderLineItemMenuException.class)
-                .isThrownBy(() -> orderValidator.validate(order));
-    }
-
-    @DisplayName("OrderLineItem 생성 후 주문 정보에 할당할 수 있다.")
-    @Test
-    void create03() {
-        // given
-        OrderLineItem 주문항목 = OrderLineItem.of(메뉴.getId(), 1L);
-
-        // when
-        주문.addAllOrderLineItems(Lists.newArrayList(주문항목));
-
-        // then
-        assertThat(주문.getOrderLineItems().getReadOnlyValues()).containsExactly(주문항목);
-    }
-
-
-
-}
