--- src/test/java/kitchenpos/fixture/MenuProductFixtureFactory.java
@@ -1,14 +1,0 @@
-package kitchenpos.fixture;
-
-
-import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuProduct;
-
-public class MenuProductFixtureFactory {
-
-    private MenuProductFixtureFactory() {}
-
-    public static MenuProduct create(Long id, Menu menu, Long productId, long quantity) {
-        return MenuProduct.of(id, menu, productId, quantity);
-    }
-}
