--- src/test/java/kitchenpos/fixture/OrderLineItemFixtureFactory.java
@@ -1,13 +1,0 @@
-package kitchenpos.fixture;
-
-
-import kitchenpos.order.domain.OrderLineItem;
-
-public class OrderLineItemFixtureFactory {
-
-    private OrderLineItemFixtureFactory() {}
-
-    public static OrderLineItem create(Long menuId, long quantity) {
-        return OrderLineItem.of(menuId, quantity);
-    }
-}
