--- src/test/java/kitchenpos/fixture/OrderTableFixtureFactory.java
@@ -1,13 +1,0 @@
-package kitchenpos.fixture;
-
-
-import kitchenpos.table.domain.OrderTable;
-
-public class OrderTableFixtureFactory {
-
-    private OrderTableFixtureFactory() {}
-
-    public static OrderTable createWithGuest(boolean empty, int numberOfGuest) {
-        return OrderTable.of(empty, numberOfGuest);
-    }
-}
