--- src/test/java/kitchenpos/fixture/TableGroupFixtureFactory.java
@@ -1,13 +1,0 @@
-package kitchenpos.fixture;
-
-
-import kitchenpos.table.domain.TableGroup;
-
-public class TableGroupFixtureFactory {
-
-    private TableGroupFixtureFactory() {}
-
-    public static TableGroup create(Long id){
-        return TableGroup.from(id);
-    }
-}
